pyproject.toml
[project]
name = "vectorbt-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cupy-cuda12x>=14.0.1",
    "nvidia-cuda-nvrtc-cu12>=12.9.86",
    "nvidia-cuda-runtime-cu12>=12.9.79",
    "nvidia-cufft-cu12>=11.4.1.4",
    "pyarrow>=24.0.0",
    "vectorbt[full,rust]>=1.0.0",
]
